backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-socketio>=5.11.0
python-engineio>=4.9.0
supabase>=2.9.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
posthog>=3.7.0
asyncpg>=0.30.0
redis>=5.2.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.12
websockets>=14.0
httpx>=0.27.0
aiofiles>=24.1.0
